Refactor business users page to use TableComponent and add search and filter functionality
