+++ src/app/admin/(dashboard)/users/[id]/business/page.tsx
@@ -1,0 +1,10 @@
+import { Text } from "@mantine/core";
+import React from "react";
+
+export default function AllBusinessUsers() {
+  return (
+    <main>
+      <Text>A business user's page</Text>
+    </main>
+  );
+}
